src/billing_insurance_service/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.1
mysqlclient==2.2.4
PyJWT==2.8.0
django-filter==24.1
celery==5.3.6
redis==5.0.1
drf-yasg==1.21.7
stripe==7.11.0
requests==2.31.0
python-dateutil==2.8.2
django-money==3.3.0
django-environ==0.11.2 